Rename method, update documentation, tweak layout
