+++ app/src/main/res/values/dimens.xml
@@ -5,7 +5,7 @@
 
     <dimen name="card_margin">16dp</dimen>
     <dimen name="card_bottom_bar_margin">60dp</dimen>
-    <dimen name="card_top_margin">32dp</dimen>
+    <dimen name="card_top_margin">16dp</dimen>
     <dimen name="card_corner_radius">0dp</dimen>
     <dimen name="card_vertical_margin">4dp</dimen>
